Tests for creator.rb and security leak fixin'
It was possible to change the state of an item that didn't belong
to the user that was logged in. Fixed that by checking if item
to change really belongs to the user logged in.

Wrote test for setting items active and inactive in controller

modified:   trade/app/controllers/creator.rb
modified:   trade/test/creator_test.rb
modified:   trade/test/test_helper.rb