More links, tables and hightlight of owner comment
	modified:   trade/app/public/css/new_style.css
added tr.self_info class to be able to highlight a owners comment. What do you hthink about the color? I just didn't like yellow, so I thought a darker, more intensive green will stand out enough.

	modified:   trade/app/views/item/active.haml
Made a link to the users as well as to the items profil.

	modified:   trade/app/views/item/item.haml
Highlighting (or "greening") the owners comment.

	modified:   trade/app/views/item/my_inactive.haml
Link to items profil.

	modified:   trade/app/views/user/id.haml
Table instead of a list. Though I'm not sure if that is a good idea. That way a user looks like an item. What do you think?