implemented Wish List
...using the observer pattern. Although it is already implemented in ruby, I
did that one myself, just because I got really confused.

Here the overall changes:

Model
--------
-wish_list.rb
very simple observer class, with basic methods such as update(item), add(item) and remove(item)

-account.rb
although we said wishlist should only be for users and not for organisations, i added the
wishlist to account. Honestly: it was just easier that moment...

-item.rb
has now a attribute called observers, where the wishlists are saved. further there are the
methods that are needed for the observer pattern.
I've insered the 'notify_ovservers' only when an item is set to inactive. As from then on, the
item is no longer in any wishlist.

Controller
-----------
- item_interaction.rb
Honestly, I'm not sure if adding/removing items from wishlists is a post or a get... currently its a post.
the post URL is /item/towishlist for adding and /item/fromwishlist for removing. I'm open form beter name
succetions.

-item_sites.rb
I decided to add the wishlist to the users home page together with his active and inactive items.

Views
--------
-wish_list.haml
a lot of copy-paste from many item tables. When viewing the wishlist, its possible to buy or remove the item from the wishlist.

-active.haml
new buttom to add to wishlist. It shows only a buttom, if item is not already in the wishlist (this is only caught in view!!)
... ähm.. the "if" is extremly ugly... any better ideas?

-my_all.haml
just added the wishlist

	modified:   trade/app/controllers/item_interaction.rb
	modified:   trade/app/controllers/item_sites.rb
	modified:   trade/app/models/account.rb
	modified:   trade/app/models/item.rb
	new file:   trade/app/models/wish_list.rb
	modified:   trade/app/require.rb
	modified:   trade/app/views/item/active.haml
	modified:   trade/app/views/item/my_all.haml
	new file:   trade/app/views/item/wish_list.haml
	new file:   trade/test/ModelTests/wish_list_spec.rb